`get_puzzle_input` optionally takes a filename
This allows us to reuse `get_puzzle_input` to load other input files
from the challenge directory. For example, it's nice to store the
smaller "example" input as `test.txt` or `example.txt` and the larger
puzzle input as `input.txt`, and then be able to easily switch between
them while working on the puzzle.