2023, Day 3, Part 1, Rust
Basic Rust implementation that's still plenty fast. Optimizing it
further might be fun and educational! Possible optimizations include:

- Avoid cloning
- Use a more efficient data structure (e.g. a HashMap of row indices)
  to avoid iterating over all of the numbers every time in
  `find_adjacent_numbers()`.

My initial implementation worked on the example but failed on the
input because I was assembling a `HashSet<i32>` of `number.value`s
instead of a `HashSet<Number>`, so I was accidentally coalescing
different Numbers with the same value. `check_input.py` was written to
quickly validate that hypothesis.