Refactor Part into a struct that implements the Index trait
This is simpler and more efficient than the previous approach of using
the newtype pattern around a `HashMap<Category, i32>`. I had just
learned about the newtype pattern and so I was excited to use it, but
then I saw this approach in another person's solution on Reddit and
instantly recognized it was superior and also a good opportunity to
learn about the `Index` trait. The newtype pattern is still being used
for `Workflows`, which I think is an excellent place for it.

Inspired by https://github.com/henryiii/aoc2023/blob/main/src/bin/19.rs.