Fix formatting of code blocks nested in list items
These looked ok when rendered on github.com, but when I rendered them
locally with `markdown-preview` in Emacs, which I have configure to
use `pandoc`, the code blocks were rendered inline instead of
separately in a `<pre>`, as I intended. Adding an extra new line
between the text of the list item and the code block fixed it.