Optimize 2023, Day 4, Python solution
Runtime: 2.2s -> 0.2s

`part2` was inefficient because it was doing lots of dynamic list
operations to simulate the process of copying cards. I replaced it
with the approach that I used in my Swift solution, which tracks the
count of each card as an integer instead of as a number of entries in
a list, and so is much more efficient because it just needs to do
arithmetic on entries in a fixed size list instead of growing a
dynamically sized list.